FlaskWeb/requirements.txt
alembic==1.0.11
argh==0.26.2
astroid==2.2.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
flasgger==0.9.3
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
google-api-core==1.16.0
google-auth==1.11.2
googleapis-common-protos==1.51.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
opencensus==0.7.7
opencensus-context==0.1.1
opencensus-ext-azure==1.0.2
opencensus-ext-flask==0.7.3
pathtools==0.1.2
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
requests==2.23.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
urllib3==1.25.8
watchdog==0.9.0
Werkzeug==0.15.5
wrapt==1.11.2
yapf==0.28.0
azure.identity==1.3.0
azure-keyvault-secrets==4.1.0